Update to OpenTelemetry v1
Otel is now GA so we can upgrade to use it and the fixed APIs. This
has meant fixing the go.mod file and some oddities with the semconv
package. Also involved was updating all the uses of attribute.Any as
this has been deprecated now, mostly moving to using Stringer with
custom String methods and the attribute.String and fmt.Sprint()